Fix session management to be done exclusively server-side (thanks Danny)
