+++ static/css/output.css
@@ -1171,6 +1171,10 @@ video {
   height: 100vh;
 }
 
+.max-h-\[600px\] {
+  max-height: 600px;
+}
+
 .max-h-full {
   max-height: 100%;
 }
