+++ static/ts/data/account.ts
@@ -5,5 +5,6 @@ import * as fonts from './_fonts';
 import * as functions from './_func';
 import * as notifications from './_notifications';
 import * as sidebar from './_sidebar';
+import * as friends from './_friends';
 
-console.log(check, colors, effects, fonts, functions, notifications, sidebar);
+console.log(check, colors, effects, fonts, functions, notifications, sidebar, friends);
