+++ templates/profile/sections/_friends.html
@@ -50,7 +50,7 @@
             <div class="flex justify-center">
                 <hr class="h-32 w-1 my-8 bg-orange-500 ml-5 mt-0 mb-0">
             </div>
-            <div class="bg-transparent border border-orange-500 h-[500px] relative w-full rounded-sm">
+            <div class="bg-transparent border border-orange-500 h-auto max-h-[600px] relative w-full rounded-sm">
                 <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="#F97316" class="bi bi-circle-fill absolute -translate-y-1/2 top-0 left-1/2 z-10" viewBox="0 0 16 16">
                     <circle cx="8" cy="8" r="8"/>
                 </svg>
@@ -102,7 +102,7 @@
                         <hr class="h-px bg-gray-200 border-0 dark:bg-gray-700">
                         {% endfor %}
                     {% else %}
-                        <img src="https://beta.shoshin.moe/static/assets/no_friends.png" class="w-1/3 mx-auto">
+                        <img src="https://beta.shoshin.moe/static/assets/nofriends.png" class="w-1/2 mx-auto mb-2">
                     {% endif %}
                 </div>
                 <div id="_sho-friendRequestsDiv" class="flex flex-col gap-2 hidden">
